pyproject.toml
[project]
name = "detr-app"
version = "0.1.0"
description = "DETR (DEtection TRansformer) Streamlit Visualization Application"
authors = [
    {name = "mutomasa", email = "mutomasa@example.com"}
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "streamlit>=1.28.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.30.0",
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
    "plotly>=5.0.0",
    "pillow>=8.3.0",
    "opencv-python>=4.5.0",
    "scikit-learn>=1.0.0",
    "scipy>=1.7.0",
    "tqdm>=4.62.0",
    "requests>=2.25.0",
    "albumentations>=1.1.0",
    "tensorboard>=2.8.0",
    "wandb>=0.12.0",
    "gradio>=3.40.0",
    "timm>=1.0.19",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
    "pre-commit>=2.15.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"] 
